requirements.txt
celery==5.2.6
Flask==2.1.0
gunicorn==20.1.0
gevent==21.8.0 
requests==2.25.1
psycopg2-binary==2.8.6
cryptography
python-dateutil